chore: update delete application to POST
