+++ src/providers/application/ApplicationProvider.tsx
@@ -329,12 +329,15 @@ function ApplicationProvider({ children }: ApplicationProviderProps) {
   async function deleteApplication() {
     dispatch({ type: ApplicationActionType.LOADING });
 
-    const response = await fetch(`/api/applications/${application!.id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
+    const response = await fetch(
+      `/api/applications/${application!.id}/delete`,
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }
+    );
 
     if (!response.ok) {
       const errorResponse = await response.json();
